New temporary branch
This is a new branch to work on and share our progress. I just updated the libft file so I could push the branch.
I will push my tokenization files later on on it but you can have a look direclty on my own branch in the meantime if you want!